Org and doc pass
