+++ README.md
@@ -19,9 +19,9 @@ https://user-images.githubusercontent.com/2632925/215318129-5bab3095-a7dd-455b-a
 
 Lots of space to play in.
 
-This is a simple floating origin plugin, useful if you want to work with very, very large scales. It works with bevy's existing `f32`-based `Transform`s, which means it's largely compatible with the bevy ecosystem. The plugin positions entities within large fixed precision grids, effectively adding precision to the location of objects.
+This is a floating origin plugin, useful if you want to work with very large or very small scales. It works with bevy's existing `f32`-based `Transform`s, which means it's largely compatible with the bevy ecosystem. The plugin positions entities within large fixed precision grids, effectively adding precision to the location of objects.
 
-Additionally, you can use reference frames to nest high precision coordinate systems. For example you might want to put all entities on a planet into the same reference frame. You can then rotate this reference frame with the planet, and orbit that planet around a star.
+Additionally, you can use reference frames to nest high precision coordinate systems. For example you might want to put all entities on a planet's surface into the same reference frame. You can then rotate this reference frame with the planet, and orbit that planet around a star.
 
 The plugin is generic over a few integer types, to trade off scale and precision for memory use. Some fun numbers with a worst case precision of 0.5mm:
   - `i8`: 2,560 km = 74% of the diameter of the Moon
@@ -30,11 +30,11 @@ The plugin is generic over a few integer types, to trade off scale and precision
   - `i64`: 19.5 million light years = ~100 times the width of the milky way galaxy
   - `i128`: 3.6e+26 light years = ~3.9e+15 times the width of the observable universe
 
-From the docs: https://docs.rs/big_space/latest/big_space/struct.GridCell.html
+This can also be used for small scales. With a cell edge length of `1e-11`, and using `i128`, there is enough precision to render objects the size of quarks anywhere in the observable universe.
+
+From the docs: https://docs.rs/big_space/latest/big_space/precision/trait.GridPrecision.html
 
 # Bevy Version Support
-
-I intend to track the `main` branch of Bevy. PRs supporting this are welcome!
 
 | bevy | big_space |
 | ---- | --------- |
