+++ src/grid_cell.rs
@@ -2,16 +2,18 @@
 
 use bevy::prelude::*;
 
-use crate::precision::GridPrecision;
+use crate::*;
+
+use self::{precision::GridPrecision, reference_frame::ReferenceFrame};
 
 /// The cell index an entity within a [`crate::ReferenceFrame`]'s grid. The [`Transform`] of an
 /// entity with this component is a transformation from the center of this cell.
 ///
 /// This component adds precision to the translation of an entity's [`Transform`]. In a
-/// high-precision [`big_space`](crate) world, the position of an entity is described by a
-/// [`Transform`] *and* a [`GridCell`]. This component is the index of a cell inside a large grid
-/// defined by a reference frame, and the transform is the position of the entity relative to the
-/// center of that cell.
+/// high-precision [`BigSpace`] world, the position of an entity is described by a [`Transform`]
+/// *and* a [`GridCell`]. This component is the index of a cell inside a large grid defined by the
+/// [`ReferenceFrame`], and the transform is the position of the entity relative to the center of
+/// that cell.
 #[derive(Component, Default, Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Hash, Reflect)]
 #[reflect(Component, Default, PartialEq)]
 pub struct GridCell<P: GridPrecision> {
@@ -42,7 +44,32 @@ impl<P: GridPrecision> GridCell<P> {
         y: P::ONE,
         z: P::ONE,
     };
+
+    /// If an entity's transform translation becomes larger than the limit specified in its
+    /// [`ReferenceFrame`], it will be relocated to the nearest grid cell to reduce the size of the
+    /// transform.
+    pub fn recenter_large_transforms(
+        reference_frames: Query<&ReferenceFrame<P>>,
+        mut changed_transform: Query<(&mut Self, &mut Transform, &Parent), Changed<Transform>>,
+    ) {
+        changed_transform
+            .par_iter_mut()
+            .for_each(|(mut grid_pos, mut transform, parent)| {
+                let Ok(reference_frame) = reference_frames.get(parent.get()) else {
+                    return;
+                };
+                if transform.as_ref().translation.abs().max_element()
+                    > reference_frame.maximum_distance_from_origin()
+                {
+                    let (grid_cell_delta, translation) = reference_frame
+                        .imprecise_translation_to_grid(transform.as_ref().translation);
+                    *grid_pos += grid_cell_delta;
+                    transform.translation = translation;
+                }
+            });
+    }
 }
+
 impl<P: GridPrecision> std::ops::Add for GridCell<P> {
     type Output = GridCell<P>;
 
@@ -54,6 +81,7 @@ impl<P: GridPrecision> std::ops::Add for GridCell<P> {
         }
     }
 }
+
 impl<P: GridPrecision> std::ops::Sub for GridCell<P> {
     type Output = GridCell<P>;
 
@@ -65,6 +93,7 @@ impl<P: GridPrecision> std::ops::Sub for GridCell<P> {
         }
     }
 }
+
 impl<P: GridPrecision> std::ops::Add for &GridCell<P> {
     type Output = GridCell<P>;
 
@@ -72,6 +101,7 @@ impl<P: GridPrecision> std::ops::Add for &GridCell<P> {
         (*self).add(*rhs)
     }
 }
+
 impl<P: GridPrecision> std::ops::Sub for &GridCell<P> {
     type Output = GridCell<P>;
