+++ src/precision.rs
@@ -13,17 +13,18 @@ use bevy::reflect::Reflect;
 /// you need. Assuming you are using a grid cell edge length of 10,000 meters, and `1.0` == 1 meter,
 /// these correspond to a total usable volume of a cube with the following edge lengths:
 ///
-/// - i8: 2,560 km = 74% of the diameter of the Moon
-/// - i16: 655,350 km = 85% of the diameter of the Moon's orbit around Earth
-/// - i32: 0.0045 light years = ~4 times the width of the solar system
-/// - i64: 19.5 million light years = ~100 times the width of the milky way galaxy
-/// - i128: 3.6e+26 light years = ~3.9e+15 times the width of the observable universe
+/// - `i8`: 2,560 km = 74% of the diameter of the Moon
+/// - `i16`: 655,350 km = 85% of the diameter of the Moon's orbit around Earth
+/// - `i32`: 0.0045 light years = ~4 times the width of the solar system
+/// - `i64`: 19.5 million light years = ~100 times the width of the milky way galaxy
+/// - `i128`: 3.6e+26 light years = ~3.9e+15 times the width of the observable universe
 ///
-/// where
+/// where `usable_edge_length = 2^(integer_bits) * cell_edge_length`, resulting in a worst case
+/// precision of 0.5mm in any of these cases.
 ///
-/// `usable_edge_length = 2^(integer_bits) * cell_edge_length`
-///
-/// resulting in a worst case precision of 0.5mm in any of these cases.
+/// This can also be used for small scales. With a cell edge length of `1e-11`, and using `i128`,
+/// there is enough precision to render objects the size of quarks anywhere in the observable
+/// universe.
 ///
 /// # Note
 ///
@@ -36,8 +37,7 @@ use bevy::reflect::Reflect;
 /// ```
 ///
 /// Additionally, consider using the provided command extensions in [`crate::commands`] to
-/// completely eliminate the use of this generic.
-///
+/// completely eliminate the use of this generic, and prevent many errors.
 pub trait GridPrecision:
     Default
     + PartialEq
